+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## 4.2.0
 * Exit process when there are no input files ([#83](https://github.com/vseventer/sharp-cli/issues/83)).
+* Added `extendWith` option to `extend` command.
+* Added `ignoreIcc` option to `composite` command.
 * Added `interFrameMaxError` and `interPaletteMaxError` output options.
+* Added `lower` and `upper` options to `normalise` command.
+* Added `reuse` alias to `optimise` and `optimize` output options.
+* Updated `sharp` dependency.
 
 ## 4.1.1 (March 21, 2023)
 * Fixes bug with AVIF filename formatting ([#28](https://github.com/vseventer/sharp-cli/issues/82)).